♻️ Refactor: client 컴포넌트 표시
